+++ assets/javascripts/discourse/components/chat-message.js
@@ -752,7 +752,6 @@ export default Component.extend({
   @action
   expand() {
     this.message.set("expanded", true);
-    this.afterExpand();
   },
 
   @action
