+++ assets/javascripts/discourse/templates/components/chat-live-pane.hbs
@@ -52,7 +52,6 @@
         setReplyTo=(action "setReplyTo")
         replyMessageClicked=(action "replyMessageClicked")
         editButtonClicked=(action "editButtonClicked")
-        afterExpand=(action "decorateMessages")
         selectingMessages=selectingMessages
         onStartSelectingMessages=onStartSelectingMessages
         onSelectMessage=onSelectMessage
